moved navbar to components
